toker bucket - part 1

+++ src/main/java/com/exs/learningsessionscrudshop/controllers/ClienteController.java
@@ -2,33 +2,36 @@ package com.exs.learningsessionscrudshop.controllers;
 
 import com.exs.learningsessionscrudshop.models.Cliente;
 import com.exs.learningsessionscrudshop.services.ClienteService;
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import io.swagger.v3.oas.annotations.media.Content;
-import io.swagger.v3.oas.annotations.media.Schema;
-import lombok.AllArgsConstructor;
+import com.exs.learningsessionscrudshop.services.RateLimitingService;
+import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 @RestController
 @RequestMapping(value = "/clientes", produces = MediaType.APPLICATION_JSON_VALUE)
-@AllArgsConstructor
+@RequiredArgsConstructor
 public class ClienteController {
 
     private static final Logger log = LoggerFactory.getLogger(ClienteController.class);
     private final ClienteService clienteService;
+    private final RateLimitingService rateLimitingService;
+
+    private static final long CAPACITY = 5;
+    private static final long WINDOW_IN_SECONDS = 60; // 1 minute expressed in seconds
 
     @PostMapping
-    @Operation(summary = "Cria um novo cliente", description = "Adiciona um novo cliente ao sistema")
-    @ApiResponse(responseCode = "200", description = "Cliente criado com sucesso",
-            content = @Content(mediaType = "application/json", schema = @Schema(implementation = Cliente.class)))
-    public ResponseEntity<Cliente> createCliente(@RequestBody Cliente cliente) {
+    public ResponseEntity<?> createCliente(@RequestBody Cliente cliente) {
+        if (!rateLimitingService.allowRequisition("createCliente", CAPACITY, WINDOW_IN_SECONDS)) {
+            return tooManyRequestsResponse();
+        }
         log.info("Iniciando criação do cliente: {}", cliente.getNome());
         Cliente savedCliente = clienteService.saveCliente(cliente);
         log.info("Cliente criado com sucesso: {}", savedCliente.getClienteId());
@@ -36,10 +39,10 @@ public class ClienteController {
     }
 
     @GetMapping
-    @Operation(summary = "Lista todos os clientes", description = "Retorna uma lista de todos os clientes")
-    @ApiResponse(responseCode = "200", description = "Lista de clientes obtida com sucesso",
-            content = @Content(mediaType = "application/json", schema = @Schema(implementation = Cliente.class)))
-    public ResponseEntity<List<Cliente>> getAllClientes() {
+    public ResponseEntity<?> getAllClientes() {
+        if (!rateLimitingService.allowRequisition("getAllClientes", CAPACITY, WINDOW_IN_SECONDS)) {
+            return tooManyRequestsResponse();
+        }
         log.info("Listando todos os clientes");
         List<Cliente> clientes = clienteService.findAllClientes();
         log.info("Total de clientes listados: {}", clientes.size());
@@ -47,11 +50,10 @@ public class ClienteController {
     }
 
     @GetMapping("/{id}")
-    @Operation(summary = "Obtém um cliente pelo ID", description = "Retorna um cliente específico pelo seu ID")
-    @ApiResponse(responseCode = "200", description = "Cliente encontrado",
-            content = @Content(mediaType = "application/json", schema = @Schema(implementation = Cliente.class)))
-    @ApiResponse(responseCode = "404", description = "Cliente não encontrado")
-    public ResponseEntity<Cliente> getClienteById(@Parameter(description = "ID do cliente") @PathVariable Long id) {
+    public ResponseEntity<?> getClienteById(@PathVariable Long id) {
+        if (!rateLimitingService.allowRequisition("getClienteById", CAPACITY, WINDOW_IN_SECONDS)) {
+            return tooManyRequestsResponse();
+        }
         log.info("Buscando cliente com ID: {}", id);
         Cliente cliente = clienteService.findClienteById(id);
         if (cliente == null) {
@@ -63,27 +65,25 @@ public class ClienteController {
     }
 
     @PutMapping("/{id}")
-    @Operation(summary = "Atualiza um cliente", description = "Atualiza os dados de um cliente existente")
-    @ApiResponse(responseCode = "200", description = "Cliente atualizado com sucesso",
-            content = @Content(mediaType = "application/json", schema = @Schema(implementation = Cliente.class)))
-    @ApiResponse(responseCode = "404", description = "Cliente não encontrado")
-    public ResponseEntity<Cliente> updateCliente(@Parameter(description = "ID do cliente") @PathVariable Long id,
-                                                 @RequestBody Cliente clienteDetails) {
+    public ResponseEntity<?> updateCliente(@PathVariable Long id, @RequestBody Cliente clienteDetails) {
+        if (!rateLimitingService.allowRequisition("updateCliente", CAPACITY, WINDOW_IN_SECONDS)) {
+            return tooManyRequestsResponse();
+        }
         log.info("Atualizando cliente com ID: {}", id);
-        Cliente cliente = clienteService.updateCliente(clienteDetails);
-        if (cliente == null) {
+        Cliente updatedCliente = clienteService.updateCliente(clienteDetails);
+        if (updatedCliente == null) {
             log.warn("Falha ao atualizar o cliente com ID: {}", id);
             return ResponseEntity.notFound().build();
         }
-        log.info("Cliente com ID: {} atualizado com sucesso.", cliente.getClienteId());
-        return ResponseEntity.ok(cliente);
+        log.info("Cliente com ID: {} atualizado com sucesso.", updatedCliente.getClienteId());
+        return ResponseEntity.ok(updatedCliente);
     }
 
     @DeleteMapping("/{id}")
-    @Operation(summary = "Exclui um cliente", description = "Remove um cliente do sistema pelo seu ID")
-    @ApiResponse(responseCode = "200", description = "Cliente excluído com sucesso")
-    @ApiResponse(responseCode = "404", description = "Cliente não encontrado")
-    public ResponseEntity<Void> deleteCliente(@Parameter(description = "ID do cliente") @PathVariable Long id) {
+    public ResponseEntity<?> deleteCliente(@PathVariable Long id) {
+        if (!rateLimitingService.allowRequisition("deleteCliente", CAPACITY, WINDOW_IN_SECONDS)) {
+            return tooManyRequestsResponse();
+        }
         log.info("Excluindo cliente com ID: {}", id);
         boolean isDeleted = clienteService.deleteCliente(id);
         if (!isDeleted) {
@@ -92,5 +92,11 @@ public class ClienteController {
         }
         log.info("Cliente com ID: {} excluído com sucesso.", id);
         return ResponseEntity.ok().build();
+    }
+
+    private ResponseEntity<Object> tooManyRequestsResponse() {
+        Map<String, String> responseMessage = new HashMap<>();
+        responseMessage.put("mensagem", "Número de chamadas por minuto excedido, tente novamente em alguns instantes");
+        return ResponseEntity.status(HttpStatus.TOO_MANY_REQUESTS).body(responseMessage);
     }
 }
